refactor of Profile and Lesson with corresponding adjustments
