+++ app/src/main/java/com/github/se/project/model/lesson/Lesson.kt
@@ -1,12 +1,12 @@
-package com.android.sample.model.lesson
+package com.github.se.project.model.lesson
 
-import com.github.se.project.model.profile.TutoringSubject
+import com.github.se.project.model.profile.Subject
 
 data class Lesson(
     val id: String = "", // Unique identifier for the lesson
     val title: String = "", // Title of the lesson
     val description: String = "", // Description of the lesson
-    val subject: TutoringSubject = TutoringSubject.PHYSICS,
+    val subject: Subject = Subject.PHYSICS,
     val tutorUid: String = "", // User ID of the tutor
     val studentUid: String = "", // User ID of the student (if booked)
     val minPrice: Double = 0.0, // Price for the lesson
@@ -15,12 +15,3 @@ data class Lesson(
     val status: LessonStatus = LessonStatus.PENDING, // Status of the lesson
     val language: String = "" // todo: create a language enum ?
 )
-
-// Enum class to define lesson status
-enum class LessonStatus {
-  REQUESTED,
-  PENDING, // Lesson is pending
-  CONFIRMED, // Lesson is confirmed
-  COMPLETED, // Lesson has been completed
-  CANCELLED // Lesson has been cancelled
-}
