+++ app/src/main/java/com/github/se/project/model/lesson/LessonStatus.kt
@@ -1,0 +1,10 @@
+package com.github.se.project.model.lesson
+
+// Enum class to define lesson status
+enum class LessonStatus {
+  REQUESTED,
+  PENDING, // Lesson is pending
+  CONFIRMED, // Lesson is confirmed
+  COMPLETED, // Lesson has been completed
+  CANCELLED // Lesson has been cancelled
+}
