+++ app/src/main/java/com/github/se/project/model/profile/AcademicLevel.kt
@@ -1,0 +1,16 @@
+package com.github.se.project.model.profile
+
+/** Enum classes representing the academic level of a user. */
+enum class AcademicLevel {
+  BA1,
+  BA2,
+  BA3,
+  BA4,
+  BA5,
+  BA6,
+  MA1,
+  MA2,
+  MA3,
+  MA4,
+  PhD,
+}
