+++ app/src/main/java/com/github/se/project/model/profile/Language.kt
@@ -1,0 +1,8 @@
+package com.github.se.project.model.profile
+
+/** Enum classes representing the language of a user. */
+enum class Language {
+  FRENCH,
+  ENGLISH,
+  GERMAN,
+}
