+++ app/src/main/java/com/github/se/project/model/profile/Section.kt
@@ -1,0 +1,20 @@
+package com.github.se.project.model.profile
+
+/** Enum classes representing the section of a user. */
+enum class Section {
+  IN,
+  SC, // School of Computer and Communication Sciences (IC)
+  AR,
+  GC,
+  SIE, // School of Architecture, Civil and Environmental Engineering (ENAC)
+  SGC,
+  MA,
+  PH, // School of Basic Sciences (SB)
+  EL,
+  MX,
+  GM,
+  MT, // School of Engineering (STI)
+  SV, // School of Life Sciences (SV)
+  NX, // Neuro-X Master (IC, STI, SV)
+  SIQ, // Quantum Science and Engineering (IC, SB, STI)
+}
