+++ app/src/main/java/com/github/se/project/model/profile/Subject.kt
@@ -1,0 +1,10 @@
+package com.github.se.project.model.profile
+
+/** Enum classes representing the subject of a user. */
+enum class Subject {
+  ANALYSIS,
+  ALGEBRA,
+  PHYSICS,
+  ICC,
+  AICC
+}
